HystoryQuikSimulator is no longer Singletone

+++ RansacBot.Net5.0/RansacsRealTime/RansacsSession.cs
@@ -28,7 +28,7 @@ namespace RansacsRealTime
 			this.monkeyNFilter.NewVertex += this.vertexes.OnNewVertex;
 		}
 
-		public RansacsSession(int N)
+		public RansacsSession(double N)
 		{
 			this.vertexes = new();
 			this.monkeyNFilter = new(N);
